Move PodMonitor to separate file
