+++ tests/test_podmonitor.py
@@ -1,0 +1,17 @@
+from unittest import TestCase
+from unittest.mock import Mock, patch, call, PropertyMock
+
+
+class PodMonitorTestCase(TestCase):
+
+    def test_add(self):
+        pass
+
+    def test_remove(self):
+        pass
+
+    def test_cleanup(self):
+        pass
+
+    def test_delete_pods(self):
+        pass
